1) Changed money to long 2) Add method which return count of bills + deleting of bills 3) Made 2 methods transactional 4) fixed REST naming

+++ src/main/java/bank/atm/dto/request/AccountRequestDto.java
@@ -3,9 +3,9 @@ package bank.atm.dto.request;
 public class AccountRequestDto {
     private Long accountNumber;
     private Long userId;
-    private double money;
+    private long money;
 
-    public double getMoney() {
+    public long getMoney() {
         return money;
     }
