+++ src/main/java/bank/atm/dto/response/AccountResponseDto.java
@@ -7,5 +7,5 @@ public class AccountResponseDto {
     private Long id;
     private Long accountNumber;
     private Long userId;
-    private double money;
+    private long money;
 }
